requirements.txt
opencv-python
flask
azure-cognitiveservices-vision-face
azure.cosmos
requests
Pillow